set continuous reporting mode to controller
Upper drivers depend on continous stream of events, when we were in
reduced reporting mode, delays in reporting were likely interpreted as
finger lost event and thing didn't work correctly.

I could have also used some timer to do the reporting, but using
built-in functionality in the controller looked like a better way.

WARNING: as I haven't found any documentation for RMI4 F12 function, I'm
not really sure about the structure of Ctrl20 register values which are
being changed. It seems to work on my controller, but it might break in
unexpected ways.