Implement up to 10 fingers support
- We report more than one descriptor
Our descriptor only supports up to 5 fingers. Technically we can also make it support 7 fingers, but we cannot go beyond that because of HID descriptor size limits. So we use Hybrid mode with only 5 fingers max in the report descriptor. This is a sane value which won't cause much issues and latency with windows. We also report now that the digitizer supports 10 touch fingers.

This commit also indirectly fixes a bug with finger touch order reported by the Synatpics controller as values can be out of order.
Another issue is fixed with clearing all of the finger data in the controller. The 950 XL synatpics controller actually supports up to 11 fingers and not 10, so we want to make sure all fingers supported by the controller get cleared out.